# api/openapi.yaml
openapi: 3.0.3
info:
  title: POS System API
  description: API for Point of Sale System with user management, product management, and sales reporting
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransactionItem:
      type: object
      properties:
        product_id:
          type: integer
          format: uint
        quantity:
          type: integer
          minimum: 1
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: uint
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
        user:
          type: string
        payment_method:
          type: string
        discount:
          type: number
          format: float64
        total_price:
          type: number
          format: float64
        created_at:
          type: string
          format: date-time
    TransactionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CreateTransactionRequest:
      type: object
      required:
        - items
        - user
        - payment_method
        - total_price
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
        user:
          type: string
        payment_method:
          type: string
        discount:
          type: number
          format: float64
        total_price:
          type: number
          format: float64
...
paths:
...
  /transactions:
    get:
      summary: List all transactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
    post:
      summary: Create a new transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /transactions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint
    get:
      summary: Get a transaction by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
...